+++ asdf/_tests/test_types.py
@@ -251,7 +251,8 @@ flow_thing:
 """
     buff = helpers.yaml_to_asdf(yaml)
     with helpers.assert_no_warnings():
-        asdf.open(buff, ignore_version_mismatch=False, extensions=CustomFlowExtension())
+        with pytest.warns(AsdfDeprecationWarning, match=".*extensions.*deprecated"):
+            asdf.open(buff, ignore_version_mismatch=False, extensions=CustomFlowExtension())
 
 
 def test_versioned_writing(monkeypatch):